add mingw-w64 (32 bit)
pretty basic mingw-w64 configuration without tools.
pthread support is built, but untested.
the package is based on the binutils and gcc versions of the
system compiler, but it should be easy to use other versions.

the approach used for the compilation of the different components
was taken from musl-cross-make, it allows to use the raw gcc
from the build directory, instead of having to install it first.
having to install it would produce some issues due to mismatch of DESTDIR
and prefix until the package is symlinked into the FS.

this is the i686 version of the mingw compiler, while it's theoretically
possible to create a single multi-lib version supporting both 64
and 32 bits, it was enough of a struggle to get this version built.
making a separate 64bit version based on this recipe should be easy,
basically it just requires a different triplet and --enable-lib64
in the mingw crt config.

mingw ships a number of tools like "genidl", but they're not needed
for basic compiler/linker usage. if it turns out that some of those
are important, they can be built from the contents of the mingw
tarball and added to the package recipe at another time.