musl: bump to 1.2.5 - ATTENTION:
this release removes basename() declaration from string.h, making it
an implicit declaration in many packages that assumed that string.h
provides it - which only holds true for glibc and their gratuitously
different prototype. instead they should include libgen.h.

an implicit declaration is nasty, because on 64bit archs it leads to
pointer truncation and therefore causes crashes that are often hard
to debug. if you apply this update, your already existing packages
will continue to work fine, but when you recompile stuff you should
grep your /src/logs/build* for "implicit declaration" and if you find
an occurence, fix it using the technique i applied to pkg/qemu.

i already fixed several packages in preparation for this update, but
not every sinlge package in the repo has been build-tested.