python-setuptools: fix build error w/o _locale module
if libintl.a (provided by gettext-tiny) isn't available during python
build, it won't create the _locale module and without it fails to
detect a default encoding when writing text files.
fix is to open the file it wants to write with wb mode, and to add
gettext to python deps (follow-up commit).