replace argp-standalone with libuargp
argp-standalone is no longer maintained
(except for a fork by @ericonr - which lacks a configure script,
and so pulls in dependencies to auto*) and the version we
used so far segfaults for some valid documented inputs, in
particular those used by "ccd2cue" utility which i'm about
to add. it's sufficient to just run ccd2cue without arguments
to experience the crash.

i looked into fixing it in the existing source, but the fix isn't
straightforward and would require too much effort.
the crash has been fixed in the glibc upstream a while ago, and
it's also included in the uclibc-derived version used by libuargp.

since libuargp is also smaller and faster to build than the
glibc-derived one, it appears to be the better choice than erico's
fork - especially since my request to add a configure script has
been ignored.

note that unlike argp-standalone, libuargp by default builds both
a static and dynamic library. i disabled the latter by overriding
the ALL_LIBS variable. argp is generally only used for command
line parsing - i.e. not for library code, so it's better to just
link it into the executable than to add a dependency to an external
.so which could cause trouble if we need to switch out the argp
implementation again.