add retroarch
the datafiles used here are actually from around retroarch 1.18 release,
but they still work and are considerably smaller.

the current libretro-assets tag is 49acc074, in case one needs/wants to
update the retroarch-database* pkgs.

the cheat database isn't installed by default because it consists
of thousands of little files that sum up to 98MB with 4k cluster
size, and isn't needed for retroarch to function.
the cores aren't installed either, they have to be built separately
using the libretro-core* packages.