gcc*: turns out libelf is only required for GCC 4.5
gcc 4.5 series introduced LTO, and libelf was required to be
installed for it to build with lto support.
we assumed that didn't change, and cargo-culted the dependency on,
but it's not required for later versions. they build fine without
and lto works regardless.

this shaves off one (albeit tiny) dependency from stage1.