remove gcc630/640
we use gcc650 since a good while, but the packages where left in place
because they used to be the stage1 compiler for a while, so it was
likely they'd show up in butch.db and some butch tools don't like
packages being referenced that don't exist in the tree (printing annoying
warnings).
meanwhile enough time has passed that everybody should have updated to
gcc650. left-over refs in butch.db should be removed.