add linux-firmware-2025
after seeing the download size of > 600mb, as opposed to the 180mb
of our 2019ish package, i decided to make the current one available
using a new name.
the old package is still fine if you run sabotage on a device older
than 5 years and probably even on newer ones, unless it's an ARM
based machine, which coincidentally is also where the huge size
increase comes from.
most of the big stuff is firmware for qualcomm and other vendors
targetting ARM boards.