gcc*: prevent debug info in libgcc_eh.a
since some parts in libgcc *need* to be compiled with -g0, the authors
switched the order of user cflags and internal cflags so the internal
ones comes last, overriding the user ones, together with a comment that
what they're doing there is conceptually wrong, but...

since -g is unconditionally added to LIBGCC2_DEBUG_CFLAGS, this causes
debug info to be added to libgcc_eh.a, which defaults to the latest dwarf
version the compiler supports, despite us passing -g0 to the build.
in gcc 11's case, that latest version is dwarf 5 which our GDB doesn't
know yet.

there's an option to configure gcc so it emits dwarf2 by default
(--with-dwarf2), but even with that in use libgcc_eh.a ends up having
dwarf5 debug info, unlike e.g. libstdc++ which in that case has dwarf2
as it should. i haven't dug into why that happens.