git: use symlinks for /lib/git-core files
turns out that by default git uses hardlinks for the command "aliases"
in /lib/git-core, which are all copies of /bin/git.
that's OK for a local install, but not when you `butch pack` the package,
as tar has no concept of preserving hardlinks so the contents of the
tarball end up being > 400MB.
since the default install works just fine for local installs not involving
butch pack, we don't bump the pkgver for this case.