butch build template: add support for new extrabins directive
this one is passed per package in vars section. it's a colon-separated
list of paths that contain binaries.
some packages install binaries into non-standard paths and call them
internally via some wrappers. an example is git which places binaries
in lib/git-core.
the purpose of passing those locations is to strip them of debug info
and place those in lib/debug if

1) a debug build is done
2) butch_do_separate_debug_info is not set to false.

while at it, some of the stripping logic was refactored.