mesalib-standard: fix build error after xvmc bump
some code links to -lXvMC but doesn't add -lXv to the command,
resulting in a "DSO missing from commandline" error when looking
for the symbol XvFreeAdaptorInfo the code uses.
the older xvmc version had xv in its pkgconfig file's "Requires"
statement, so it was pulled in accidentally.
since it's mesa's code using xv and not xvmc, we fix it here.