add optional C implementation of butch-relocate
there are some packages with lots of files, like ca-certificates,
and there there's a noticable delay when running butch relocate/
relink. linking all packages in an entire "core" rootfs into place
takes almost 20 seconds, because there's quite some overhead
when doing this in shell.
this new C version finishes instantly, however it only saw basic
testing so it's not recommended to use it yet - except for testing.
the code is basically a 1-to-1 translation of the shell script,
but there are many subtle differences between shell commands and
their C syscall equivalents, so it's expected that some issues
in cornercases will show up when using it.
the use of this program will always stay optional, since one of
butch's nicest features is that one can run it directly from a
git checkout, and the speed of the shellscript is sufficient
for regular use.
future changes to the business logic in butch-relocate will need
to be applied to both the C and shell version.