fix deps in a couple packages
note that often the default dependencies are for a "minimally usable"
configuration. for example for mpv and sdl2 one might also want
mesalib for openGL acceleration.
we don't make that default though since users on low-end devices
might only need a framebuffer and mesalib is one of the heaviest
packages in the system.