butch-genfilelist: use sort -u
all users of the script currently do this themselves.
move it to the script so it can be handled in one place.
most of the added complexity is due to that -o pipefail was only
added to posix in 2024, so we can't rely on it being there for
few years to come.