build template: make sure relocation succeeds
i encountered silent breakage using CONFIG=./config butch install XXX
because the last part of the build template cds into the install dir
for various tasks, which invalidates the relative CONFIG path and caused
the sourcing to fail.
i'm not clear on why this caused "butch relocate" to bail out since
there's no "set -e" in it.

this is a fallout from the change from "butch-relocate" to "butch relocate"
call, as the former does not go through the wrapper which is in charge of
CONFIG handling.
this was changed because of the addition of the new butch-relocate-c
C implementation that's a lot faster and is used if the main butch wrapper
detects its presence.

however, this is conceptually wrong, only the butch wrapper should deal
with CONFIG, and all other butch tasks started from there should use the
commands directly from BINDIR, without going through the wrapper again.
instead, they should check the presence and validity of variables supposed
to be sourced from CONFIG.
that means we need to revert that change, and make it so butch-relocate
does the detection of whether the C binary exists.

however it's not a bad idea to check whether butch-relocate fails anyway.
an additional safeguard could be to turn relative CONFIG paths into
absolute.