Revert "butch: if there's a butch-relocate-c in BINDIR, use it"
This reverts commit a0a94166ef626eac5cfa5a8a6754f6142271e587.

we'll move the functionality to butch-relocate itself, so all
users of butch-relocate can take advantage of it, without having
to go through the wrapper (again).
the latter can cause problems due to a relative path used in CONFIG
no longer pointing to the right place in case the current working
directory was changed.