butch: bump version number and wire up the new butch-relocate-c
the new, much faster C implementation of butch-relocate is now
either installed via sabotage-core, or automatically compiled on
demand when running butch relink/relocate, if possible.
a couple of safeguards were built in so that butch can gracefully
fall back to the shell implementation, if the compilation isn't
possible or doesn't succeed for other reasons.