butch-checksum: fix awk detection code
the assignments for the fallback awk impls didn't use necessary quotes.
also the >&2 redirects caused the awk xor result tests to fail
in set -x mode (for some reason, the so-redirected stderr contained
the shell stderr output with the executed command line, not the awk
output).

also simplify the filelist iteration code analoguous to butch-unlink.