README: add info how to download stage1 packages in advance
all packages that require a https download (such as musl-cross-make-tarball)
required to build stage1, currently fail to download since busybox was bumped
to 1.33.x and lost the ability to fetch https pages.
once stage1 is built inside the chroot, curl is available and further downloads
will take advantage of it.

this issue slipped through my radar since the tarballs were already downloaded
on the build machine.